Hold off from potentially dangerous action for now
If we choose to erase shared state(s) upon the application stopping and
there are still hundreds (if not thousands) of references to it (e.g.
open TLS connections), a sudden spike in memory usage could occur.

Let's briefly consider a shared state occupying ~74 KiB of data (a
realistic amount, following a measurement of my own using
`erts_debug:flat_size/1`.)

Multiplying this a thousand fold will suddenly increase occupied
memory by up to ~74 MiB[*] (not counting the overhead), spread
over a thousand independent and newly created objects. Not good.

On the other hand, by not erasing the shared state, at worse we're
surrendering those extra 74 KiB of memory as "lost". Not ideal, but not
really a problem nowadays, either.

Let's go with the latter, then.

[*]: A simplification, since only part of the shared state is kept (the
list of certification authorities.)